separate viewers into their own files
