+++ src/hubbleds/viewers/__init__.py
@@ -1,2 +1,3 @@
-from .viewers import *
 from .hubble_dotplot import *
+from .hubble_scatter_viewer import *
+from .hubble_fit_viewer import *
