+++ src/hubbleds/viewers/hubble_fit_viewer.py
@@ -1,22 +1,12 @@
 from echo import delay_callback, add_callback
 from glue_plotly.viewers.scatter import PlotlyScatterView
-from cosmicds.viewers import CDSScatterViewerState
+from .hubble_scatter_viewer import HubbleScatterViewerState
 from cosmicds.viewers import cds_viewer
 
 __all__ = [
-    "HubbleScatterViewerState", "HubbleFitViewerState",
-    "HubbleFitView", "HubbleScatterView",
+    "HubbleFitView",
+    "HubbleFitLayerView",
 ]
-
-
-class HubbleScatterViewerState(CDSScatterViewerState):
-
-    def reset_limits(self, visible_only=True):
-        with delay_callback(self, 'x_min', 'x_max', 'y_min', 'y_max'):
-            super().reset_limits(visible_only=visible_only)
-            self.x_min = min(self.x_min, 0) if self.x_min is not None else 0
-            self.y_min = min(self.y_min, 0) if self.y_min is not None else 0
-
 
 class HubbleFitViewerState(HubbleScatterViewerState):
     
@@ -27,6 +17,7 @@ class HubbleFitViewerState(HubbleScatterViewerState):
                 self.x_max = 1.1 * self.x_max
             if self.y_max is not None:
                 self.y_max = 1.1 * self.y_max
+
 
 
 HubbleFitView = cds_viewer(
@@ -50,16 +41,4 @@ HubbleFitLayerView = cds_viewer(
     ],
     label='Layer View',
     state_cls=HubbleFitViewerState
-)
-
-HubbleScatterView = cds_viewer(
-    PlotlyScatterView,
-    name="HubbleScatterView",
-    viewer_tools=[
-        'plotly:home',
-        'plotly:zoom',
-        'hubble:linefit'
-    ],
-    label='Scatter View',
-    state_cls=HubbleScatterViewerState
 )
